backups/questcity-backend-backup-20250908_114522/pyproject.toml
[tool.poetry]
name = "quests"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
alembic = "^1.13.3"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.0"
aioinject = "^0.35.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.12"
result = "^0.17.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
cryptography = "^43.0.3"
python-jose = "^3.3.0"
pyyaml = "^6.0.2"
fastapi-mail = "^1.4.1"
httpx = "^0.27.2"
itsdangerous = "^2.2.0"
fastapi-pagination = "^0.12.31"
fastapi-filter = {extras = ["sqlalchemy"], version = "^2.0.0"}
python-json-logger = "^3.2.1"
aioboto3 = "^13.4.0"
python-magic = "^0.4.27"
validators = "^0.34.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.38"}
pillow = "^11.1.0"
websockets = "^15.0.1"
structlog = "^25.4.0"
requests = "^2.32.4"
python-magic-bin = "^0.4.14"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.0"
isort = "^5.13.2"
mypy = "^1.12.0"
pyclean = "^3.0.0"
pre-commit = "^4.2.0"
types-requests = "^2.32.4.20250611"
types-pyyaml = "^6.0.12.20250516"
types-python-dateutil = "^2.9.0.20250708"
types-setuptools = "^80.9.0.20250529"
types-pillow = "^10.2.0.20240822"
types-redis = "^4.6.0.20241004"
types-cryptography = "^3.3.23.2"
boto3-stubs = "^1.39.14"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"